+++ lib/constantes.dart
@@ -32,6 +32,9 @@ class MainProject {
   /// Endpoint = /documents
   static const String documentsEndPoint = '/documents';
 
+  /// Endpoint = /documents
+  static const String services = '/proposed-services';
+
   /// Endpoint = '/documents/upload'
   static const String documentsUploadEndPoint = '/documents/upload';
 
@@ -80,6 +83,9 @@ class MenuText {
 
   /// Value = "authentification"
   static String auth = 'authentification';
+
+  /// Value = "Services"
+  static String services = 'Services';
 }
 
 class ErrorMessage {