+++ lib/models/proposed_service/proposed_service_model.freezed.dart
@@ -1,0 +1,316 @@
+// coverage:ignore-file
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
+
+part of 'proposed_service_model.dart';
+
+// **************************************************************************
+// FreezedGenerator
+// **************************************************************************
+
+T _$identity<T>(T value) => value;
+
+final _privateConstructorUsedError = UnsupportedError(
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#adding-getters-and-methods-to-our-models');
+
+ProposedServiceModel _$ProposedServiceModelFromJson(Map<String, dynamic> json) {
+  return _ProposedServiceModel.fromJson(json);
+}
+
+/// @nodoc
+mixin _$ProposedServiceModel {
+  String get title => throw _privateConstructorUsedError;
+  String get content => throw _privateConstructorUsedError;
+  bool get isPublished => throw _privateConstructorUsedError;
+  String get createdAt => throw _privateConstructorUsedError;
+  String get updatedAt => throw _privateConstructorUsedError;
+  String? get icon => throw _privateConstructorUsedError;
+  String? get deletedAt => throw _privateConstructorUsedError;
+  int? get id => throw _privateConstructorUsedError;
+
+  /// Serializes this ProposedServiceModel to a JSON map.
+  Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
+
+  /// Create a copy of ProposedServiceModel
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  $ProposedServiceModelCopyWith<ProposedServiceModel> get copyWith =>
+      throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+abstract class $ProposedServiceModelCopyWith<$Res> {
+  factory $ProposedServiceModelCopyWith(ProposedServiceModel value,
+          $Res Function(ProposedServiceModel) then) =
+      _$ProposedServiceModelCopyWithImpl<$Res, ProposedServiceModel>;
+  @useResult
+  $Res call(
+      {String title,
+      String content,
+      bool isPublished,
+      String createdAt,
+      String updatedAt,
+      String? icon,
+      String? deletedAt,
+      int? id});
+}
+
+/// @nodoc
+class _$ProposedServiceModelCopyWithImpl<$Res,
+        $Val extends ProposedServiceModel>
+    implements $ProposedServiceModelCopyWith<$Res> {
+  _$ProposedServiceModelCopyWithImpl(this._value, this._then);
+
+  // ignore: unused_field
+  final $Val _value;
+  // ignore: unused_field
+  final $Res Function($Val) _then;
+
+  /// Create a copy of ProposedServiceModel
+  /// with the given fields replaced by the non-null parameter values.
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? title = null,
+    Object? content = null,
+    Object? isPublished = null,
+    Object? createdAt = null,
+    Object? updatedAt = null,
+    Object? icon = freezed,
+    Object? deletedAt = freezed,
+    Object? id = freezed,
+  }) {
+    return _then(_value.copyWith(
+      title: null == title
+          ? _value.title
+          : title // ignore: cast_nullable_to_non_nullable
+              as String,
+      content: null == content
+          ? _value.content
+          : content // ignore: cast_nullable_to_non_nullable
+              as String,
+      isPublished: null == isPublished
+          ? _value.isPublished
+          : isPublished // ignore: cast_nullable_to_non_nullable
+              as bool,
+      createdAt: null == createdAt
+          ? _value.createdAt
+          : createdAt // ignore: cast_nullable_to_non_nullable
+              as String,
+      updatedAt: null == updatedAt
+          ? _value.updatedAt
+          : updatedAt // ignore: cast_nullable_to_non_nullable
+              as String,
+      icon: freezed == icon
+          ? _value.icon
+          : icon // ignore: cast_nullable_to_non_nullable
+              as String?,
+      deletedAt: freezed == deletedAt
+          ? _value.deletedAt
+          : deletedAt // ignore: cast_nullable_to_non_nullable
+              as String?,
+      id: freezed == id
+          ? _value.id
+          : id // ignore: cast_nullable_to_non_nullable
+              as int?,
+    ) as $Val);
+  }
+}
+
+/// @nodoc
+abstract class _$$ProposedServiceModelImplCopyWith<$Res>
+    implements $ProposedServiceModelCopyWith<$Res> {
+  factory _$$ProposedServiceModelImplCopyWith(_$ProposedServiceModelImpl value,
+          $Res Function(_$ProposedServiceModelImpl) then) =
+      __$$ProposedServiceModelImplCopyWithImpl<$Res>;
+  @override
+  @useResult
+  $Res call(
+      {String title,
+      String content,
+      bool isPublished,
+      String createdAt,
+      String updatedAt,
+      String? icon,
+      String? deletedAt,
+      int? id});
+}
+
+/// @nodoc
+class __$$ProposedServiceModelImplCopyWithImpl<$Res>
+    extends _$ProposedServiceModelCopyWithImpl<$Res, _$ProposedServiceModelImpl>
+    implements _$$ProposedServiceModelImplCopyWith<$Res> {
+  __$$ProposedServiceModelImplCopyWithImpl(_$ProposedServiceModelImpl _value,
+      $Res Function(_$ProposedServiceModelImpl) _then)
+      : super(_value, _then);
+
+  /// Create a copy of ProposedServiceModel
+  /// with the given fields replaced by the non-null parameter values.
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? title = null,
+    Object? content = null,
+    Object? isPublished = null,
+    Object? createdAt = null,
+    Object? updatedAt = null,
+    Object? icon = freezed,
+    Object? deletedAt = freezed,
+    Object? id = freezed,
+  }) {
+    return _then(_$ProposedServiceModelImpl(
+      title: null == title
+          ? _value.title
+          : title // ignore: cast_nullable_to_non_nullable
+              as String,
+      content: null == content
+          ? _value.content
+          : content // ignore: cast_nullable_to_non_nullable
+              as String,
+      isPublished: null == isPublished
+          ? _value.isPublished
+          : isPublished // ignore: cast_nullable_to_non_nullable
+              as bool,
+      createdAt: null == createdAt
+          ? _value.createdAt
+          : createdAt // ignore: cast_nullable_to_non_nullable
+              as String,
+      updatedAt: null == updatedAt
+          ? _value.updatedAt
+          : updatedAt // ignore: cast_nullable_to_non_nullable
+              as String,
+      icon: freezed == icon
+          ? _value.icon
+          : icon // ignore: cast_nullable_to_non_nullable
+              as String?,
+      deletedAt: freezed == deletedAt
+          ? _value.deletedAt
+          : deletedAt // ignore: cast_nullable_to_non_nullable
+              as String?,
+      id: freezed == id
+          ? _value.id
+          : id // ignore: cast_nullable_to_non_nullable
+              as int?,
+    ));
+  }
+}
+
+/// @nodoc
+@JsonSerializable()
+class _$ProposedServiceModelImpl implements _ProposedServiceModel {
+  const _$ProposedServiceModelImpl(
+      {required this.title,
+      required this.content,
+      required this.isPublished,
+      required this.createdAt,
+      required this.updatedAt,
+      this.icon,
+      this.deletedAt,
+      this.id});
+
+  factory _$ProposedServiceModelImpl.fromJson(Map<String, dynamic> json) =>
+      _$$ProposedServiceModelImplFromJson(json);
+
+  @override
+  final String title;
+  @override
+  final String content;
+  @override
+  final bool isPublished;
+  @override
+  final String createdAt;
+  @override
+  final String updatedAt;
+  @override
+  final String? icon;
+  @override
+  final String? deletedAt;
+  @override
+  final int? id;
+
+  @override
+  String toString() {
+    return 'ProposedServiceModel(title: $title, content: $content, isPublished: $isPublished, createdAt: $createdAt, updatedAt: $updatedAt, icon: $icon, deletedAt: $deletedAt, id: $id)';
+  }
+
+  @override
+  bool operator ==(Object other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType &&
+            other is _$ProposedServiceModelImpl &&
+            (identical(other.title, title) || other.title == title) &&
+            (identical(other.content, content) || other.content == content) &&
+            (identical(other.isPublished, isPublished) ||
+                other.isPublished == isPublished) &&
+            (identical(other.createdAt, createdAt) ||
+                other.createdAt == createdAt) &&
+            (identical(other.updatedAt, updatedAt) ||
+                other.updatedAt == updatedAt) &&
+            (identical(other.icon, icon) || other.icon == icon) &&
+            (identical(other.deletedAt, deletedAt) ||
+                other.deletedAt == deletedAt) &&
+            (identical(other.id, id) || other.id == id));
+  }
+
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  @override
+  int get hashCode => Object.hash(runtimeType, title, content, isPublished,
+      createdAt, updatedAt, icon, deletedAt, id);
+
+  /// Create a copy of ProposedServiceModel
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  @override
+  @pragma('vm:prefer-inline')
+  _$$ProposedServiceModelImplCopyWith<_$ProposedServiceModelImpl>
+      get copyWith =>
+          __$$ProposedServiceModelImplCopyWithImpl<_$ProposedServiceModelImpl>(
+              this, _$identity);
+
+  @override
+  Map<String, dynamic> toJson() {
+    return _$$ProposedServiceModelImplToJson(
+      this,
+    );
+  }
+}
+
+abstract class _ProposedServiceModel implements ProposedServiceModel {
+  const factory _ProposedServiceModel(
+      {required final String title,
+      required final String content,
+      required final bool isPublished,
+      required final String createdAt,
+      required final String updatedAt,
+      final String? icon,
+      final String? deletedAt,
+      final int? id}) = _$ProposedServiceModelImpl;
+
+  factory _ProposedServiceModel.fromJson(Map<String, dynamic> json) =
+      _$ProposedServiceModelImpl.fromJson;
+
+  @override
+  String get title;
+  @override
+  String get content;
+  @override
+  bool get isPublished;
+  @override
+  String get createdAt;
+  @override
+  String get updatedAt;
+  @override
+  String? get icon;
+  @override
+  String? get deletedAt;
+  @override
+  int? get id;
+
+  /// Create a copy of ProposedServiceModel
+  /// with the given fields replaced by the non-null parameter values.
+  @override
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  _$$ProposedServiceModelImplCopyWith<_$ProposedServiceModelImpl>
+      get copyWith => throw _privateConstructorUsedError;
+}